requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.1
h11==0.14.0
idna==3.7
motor==3.5.1
numpy==1.24.4
pandas==2.0.3
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
